# Switch all resize tests to bicubic interpolation