This switches all resize tests to use bicubic interpolation. See the relevant discussion here: https://github.com/bleroy/core-imaging-playground/pull/15#discussion_r286902466.

I've also slightly adapted the NetVips resize benchmark, it's now using the convenient [`vips_resize`](https://libvips.github.io/libvips/API/current/libvips-resample.html#vips-resize) instead of the low-level [`vips_reduce`](https://libvips.github.io/libvips/API/current/libvips-resample.html#vips-reduce) operation. The [`CopyMemory()`](https://libvips.github.io/libvips/API/current/VipsImage.html#vips-image-copy-memory) operations introduced in PR #16 doesn't seem quite right to me. I changed it to [`WriteToMemory()`](https://libvips.github.io/libvips/API/current/VipsImage.html#vips-image-write-to-memory) at the end of the chain to have libvips actually process pixels.

<details><summary>Benchmark results on Linux</summary>

**Runtime information**
```ini

BenchmarkDotNet=v0.11.5, OS=fedora 30
Intel Core i5-8600K CPU 3.60GHz (Coffee Lake), 1 CPU, 6 logical and 6 physical cores
.NET Core SDK=3.0.100
  [Host]            : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT
  .Net Core 3.0 CLI : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT

Job=.Net Core 3.0 CLI  Toolchain=.NET Core 3.0  IterationCount=5  
LaunchCount=1  WarmupCount=5  
```

**Before**
```
|                    Method |      Mean |     Error |    StdDev | Ratio |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |----------:|----------:|----------:|------:|---------:|------:|------:|----------:|
|   'System.Drawing Resize' | 50.686 ms | 0.3015 ms | 0.0783 ms |  1.00 |        - |     - |     - |     128 B |
|       'ImageSharp Resize' |  5.599 ms | 0.1256 ms | 0.0326 ms |  0.11 |        - |     - |     - |    7440 B |
|      'ImageMagick Resize' | 31.789 ms | 0.2118 ms | 0.0550 ms |  0.63 |        - |     - |     - |    5656 B |
|        'FreeImage Resize' |  7.302 ms | 0.0637 ms | 0.0165 ms |  0.14 | 500.0000 |     - |     - |     136 B |
| 'SkiaSharp Canvas Resize' |  1.344 ms | 0.0053 ms | 0.0014 ms |  0.03 |        - |     - |     - |     636 B |
| 'SkiaSharp Bitmap Resize' |  1.326 ms | 0.0013 ms | 0.0003 ms |  0.03 |        - |     - |     - |     842 B |
|          'NetVips Resize' |  2.727 ms | 0.0116 ms | 0.0030 ms |  0.05 |        - |     - |     - |    5320 B |
```

**After**
```
|                    Method |        Mean |      Error |      StdDev | Ratio |    Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------------- |------------:|-----------:|------------:|------:|---------:|-------:|------:|----------:|
|   'System.Drawing Resize' | 50,828.5 us | 617.288 us | 160.3078 us |  1.00 |        - |      - |     - |     128 B |
|       'ImageSharp Resize' |  5,708.7 us |  82.080 us |  21.3159 us |  0.11 |        - |      - |     - |    7440 B |
|      'ImageMagick Resize' | 31,754.7 us | 254.287 us |  66.0376 us |  0.62 |        - |      - |     - |    5656 B |
|        'FreeImage Resize' |  7,286.5 us |  28.773 us |   7.4723 us |  0.14 | 500.0000 |      - |     - |     136 B |
| 'SkiaSharp Canvas Resize' |  1,338.3 us |   6.414 us |   1.6658 us |  0.03 |        - |      - |     - |     636 B |
| 'SkiaSharp Bitmap Resize' |    628.7 us |   2.284 us |   0.5933 us |  0.01 |        - |      - |     - |     841 B |
|          'NetVips Resize' |  1,107.2 us |  52.898 us |  13.7374 us |  0.02 |   3.9063 | 1.9531 |     - |   20360 B |
```
</details>


<details><summary>Benchmark results on Windows</summary>

**Runtime information**
```ini
BenchmarkDotNet=v0.11.5, OS=Windows 10.0.18362
Intel Core i5-8600K CPU 3.60GHz (Coffee Lake), 1 CPU, 6 logical and 6 physical cores
.NET Core SDK=3.0.100
  [Host]            : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT
  .Net Core 3.0 CLI : .NET Core 3.0.0 (CoreCLR 4.700.19.46205, CoreFX 4.700.19.46214), 64bit RyuJIT

Job=.Net Core 3.0 CLI  Toolchain=.NET Core 3.0  IterationCount=5  
LaunchCount=1  WarmupCount=5  
```

**Before**
```
|                    Method |      Mean |     Error |    StdDev | Ratio | RatioSD |    Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |----------:|----------:|----------:|------:|--------:|---------:|------:|------:|----------:|
|   'System.Drawing Resize' | 10.369 ms | 0.1315 ms | 0.0341 ms |  1.00 |    0.00 |        - |     - |     - |     136 B |
|       'ImageSharp Resize' |  5.676 ms | 0.0196 ms | 0.0051 ms |  0.55 |    0.00 |        - |     - |     - |    7440 B |
|      'ImageMagick Resize' | 43.927 ms | 0.2877 ms | 0.0747 ms |  4.24 |    0.02 |        - |     - |     - |    5656 B |
|        'FreeImage Resize' |  7.884 ms | 0.0350 ms | 0.0091 ms |  0.76 |    0.00 | 500.0000 |     - |     - |     136 B |
|      'MagicScaler Resize' |  3.360 ms | 0.0777 ms | 0.0202 ms |  0.32 |    0.00 |  11.7188 |     - |     - |   59525 B |
| 'SkiaSharp Canvas Resize' |  4.958 ms | 0.0235 ms | 0.0061 ms |  0.48 |    0.00 |        - |     - |     - |     653 B |
| 'SkiaSharp Bitmap Resize' |  4.940 ms | 0.0222 ms | 0.0058 ms |  0.48 |    0.00 |        - |     - |     - |     848 B |
|          'NetVips Resize' |  4.901 ms | 0.1076 ms | 0.0279 ms |  0.47 |    0.00 |        - |     - |     - |    5320 B |
```

**After**
```
|                    Method |      Mean |     Error |    StdDev | Ratio |    Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------------------------- |----------:|----------:|----------:|------:|---------:|-------:|------:|----------:|
|   'System.Drawing Resize' | 10.344 ms | 0.1375 ms | 0.0357 ms |  1.00 |        - |      - |     - |     136 B |
|       'ImageSharp Resize' |  5.644 ms | 0.1020 ms | 0.0265 ms |  0.55 |        - |      - |     - |    7440 B |
|      'ImageMagick Resize' | 43.887 ms | 0.4111 ms | 0.1068 ms |  4.24 |        - |      - |     - |    5656 B |
|        'FreeImage Resize' |  7.887 ms | 0.0708 ms | 0.0184 ms |  0.76 | 500.0000 |      - |     - |     136 B |
|      'MagicScaler Resize' |  3.351 ms | 0.0059 ms | 0.0015 ms |  0.32 |  11.7188 |      - |     - |   59525 B |
| 'SkiaSharp Canvas Resize' |  4.973 ms | 0.0114 ms | 0.0030 ms |  0.48 |        - |      - |     - |     650 B |
| 'SkiaSharp Bitmap Resize' |  2.461 ms | 0.3670 ms | 0.0953 ms |  0.24 |        - |      - |     - |     844 B |
|          'NetVips Resize' |  1.157 ms | 0.0052 ms | 0.0014 ms |  0.11 |   3.9063 | 1.9531 |     - |   20360 B |
```
</details>

/cc @jcupitt, because he might also be interested in the numbers above.