@jcupitt 2019-10-15
Huh the benchmarks look nice Kleis. It's great you get the same time for NetVips on linux and win now.