@kleisauke 2019-10-15
@jcupitt This is the 'simple' resize benchmark results, here are the 'Load, Resize, Save' benchmark results:

<details><summary>'Load, Resize, Save' benchmark results on Linux</summary>

```
|                                Method |      Mean |     Error |    StdDev | Ratio |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|-------------------------------------- |----------:|----------:|----------:|------:|----------:|------:|------:|-----------:|
|   'System.Drawing Load, Resize, Save' | 252.97 ms | 0.3161 ms | 0.0821 ms |  1.00 |         - |     - |     - |    9.47 KB |
|       'ImageSharp Load, Resize, Save' | 255.27 ms | 0.6467 ms | 0.1680 ms |  1.01 |         - |     - |     - | 1106.15 KB |
|      'ImageMagick Load, Resize, Save' | 349.53 ms | 2.9678 ms | 0.7707 ms |  1.38 |         - |     - |     - |    55.5 KB |
|        'ImageFree Load, Resize, Save' | 204.31 ms | 1.0210 ms | 0.2651 ms |  0.81 | 6000.0000 |     - |     - |   98.59 KB |
| 'SkiaSharp Canvas Load, Resize, Save' | 130.89 ms | 3.5666 ms | 0.9262 ms |  0.52 |         - |     - |     - |  189.43 KB |
| 'SkiaSharp Bitmap Load, Resize, Save' | 130.21 ms | 0.5455 ms | 0.1417 ms |  0.51 |         - |     - |     - |  185.57 KB |
|          'NetVips Load, Resize, Save' |  97.06 ms | 0.7699 ms | 0.1999 ms |  0.38 |         - |     - |     - |   92.27 KB |
```
</details>

<details><summary>'Load, Resize, Save' benchmark results on Windows</summary>

```
|                                Method |      Mean |    Error |    StdDev | Ratio |     Gen 0 | Gen 1 | Gen 2 |  Allocated |
|-------------------------------------- |----------:|---------:|----------:|------:|----------:|------:|------:|-----------:|
|   'System.Drawing Load, Resize, Save' | 376.67 ms | 2.716 ms | 0.7054 ms |  1.00 |         - |     - |     - |   10.02 KB |
|       'ImageSharp Load, Resize, Save' | 241.92 ms | 8.708 ms | 2.2616 ms |  0.64 |         - |     - |     - | 1103.41 KB |
|      'ImageMagick Load, Resize, Save' | 399.66 ms | 6.163 ms | 1.6004 ms |  1.06 |         - |     - |     - |   55.66 KB |
|        'ImageFree Load, Resize, Save' | 259.29 ms | 1.085 ms | 0.2818 ms |  0.69 | 6000.0000 |     - |     - |   95.67 KB |
|      'MagicScaler Load, Resize, Save' |  88.93 ms | 8.158 ms | 2.1187 ms |  0.24 |         - |     - |     - |  343.49 KB |
| 'SkiaSharp Canvas Load, Resize, Save' | 186.92 ms | 2.022 ms | 0.5250 ms |  0.50 |         - |     - |     - |  185.47 KB |
| 'SkiaSharp Bitmap Load, Resize, Save' | 186.67 ms | 2.204 ms | 0.5723 ms |  0.50 |         - |     - |     - |   182.3 KB |
|          'NetVips Load, Resize, Save' | 146.28 ms | 3.662 ms | 0.9510 ms |  0.39 |         - |     - |     - |   92.42 KB |
```
</details>

NetVips is still about 1.5 times slower on Windows than on Linux, but maybe we can't do much about it. The decrease in performance on Windows is also noticeable on the other image processing benchmarks.

Note that this is without the 'fair-competition patch' (YCbCr vs RGB – Catmull-Rom vs Lanczos3) mentioned in #15, that's why MagicScaler differs from the rest.