@jcupitt 2019-10-15
Ah, interesting, I wondered if file IO was in there too.

I did notice a small performance issue yesterday with libpng read in libvips -- it's doing many more syscalls than it should. I don't suppose this is using libpng?