@kleisauke 2019-10-15
Only jpeg images are tested with this benchmark, so this is using libjpeg-turbo. Regarding libpng, you might also be interested in this issue: https://github.com/lovell/sharp-libvips/issues/25. See the relevant benchmarks: https://gist.github.com/kleisauke/879e3075a675c6f945dff3772d2ef0a3.

Hopefully [libspng](https://libspng.org/) will stabilize soon, it should be up to 35% faster than libpng(!).