@bleroy 2019-10-15
Thanks for the contribution. I'm happy to take most of it, but I'd like @JimBobSquarePants ' opinion on the bicubic change to ImageSharp before I merge.