@saucecontrol 2019-10-15
Based on https://github.com/mono/SkiaSharp/issues/520#issuecomment-442923077 and the fact that the timing numbers and image quality are the same, it seems that there is no longer any reason to have two versions of each Skia test.  They appear to be doing the same thing now.