@kleisauke 2019-10-15
Hi @saucecontrol,

Thank you for your comments and code suggestions! I guess the shortcut what you're seeing is due to the operation cache. Every time you call an operation, libvips searches the cache for a previous call to the same operation with the same arguments. If it finds a match, you get the previous result again. You can disable this behavior with:
```csharp
NetVips.CacheSetMax(0);
```

You could also try to disable the run-time vector code generation and set the the number of worker threads to 1:
```csharp
NetVips.VectorSet(false); 
NetVips.ConcurrencySet(1);
```

I should not call it 'fair-competition patch', sorry about that. I just wanted to explain why MagicScaler is a bit faster on Windows with the 'Load, Resize, Save' benchmark. The reason that libvips is slower here is because of the threading, thumbnailing to 150x150 pixels can be done quite quickly, so a large thread pool could slow down processing overall.

Regarding the correct output in the load/resize/save tests, could you try this patch?:

```diff
diff --git a/NetCore/LoadResizeSave.cs b/NetCore/LoadResizeSave.cs
index 1111111..2222222 100644
--- a/NetCore/LoadResizeSave.cs
+++ b/NetCore/LoadResizeSave.cs
@@ -187,9 +187,6 @@ namespace ImageProcessing
                 // Resize it to fit a 150x150 square
                 image.Resize(ThumbnailSize, ThumbnailSize);
 
-                // Reduce the size of the file
-                image.Strip();
-
                 // Set the quality
                 image.Quality = Quality;
 
@@ -331,7 +328,7 @@ namespace ImageProcessing
             using (var thumb = NetVipsImage.Thumbnail(input, ThumbnailSize, ThumbnailSize))
             {
                 // Save the results
-                thumb.Jpegsave(OutputPath(input, NetVips), q: Quality, strip: true);
+                thumb.Jpegsave(OutputPath(input, NetVips), q: Quality);
             }
         }
     }
```

Note that this also increases the file size, because EXIF data, IPTC data, or XMP metadata are now also being saved.