@jcupitt 2019-10-15
This is an interesting discussion, thank you!

Re. the operation cache: all libvips operations are functional, that is, they have no side-effects. This means you can replace the result of any operation with the result of a previous call with the same arguments. By default, libvips aims to keep about 100mb of cached operations.

This can give misleading benchmark results if you loop several times over the same image. I usually try to have N distinct input images rather than looping N times over the same image.